fix: uses nullish coalescing operator for defaults
Replaces logical OR operators with nullish coalescing operators (??) for providing default values in titles, descriptions, and other text elements across the application. This ensures that default values are only used when a variable is null or undefined, and not when it's an empty string or other falsy value.

Renames `.ts` files to `.tsx` for OG image routes to align with their usage of JSX.

Also, fixes type definition for `generateMetadata` props in governance app.