Improves Vault secrets retrieval
Dynamically determines the environment name for retrieving database secrets from Vault.

This allows the action to retrieve the correct secrets based on whether the workflow is triggered by a pull request or a direct push to a branch. For pull requests, it checks the base branch to determine if it's targeting `main` (production) or another branch (development).  For direct pushes, it checks the ref name.