feat(verifier): simple target collapser (#409)
This change adds a simple target duplicate finder. We work on top of slices
by comparing distances of the slice indices. This provides a reliable way to find
_almost_ identical predictions. This finds changes like missing dots and whitespaces
or broken target slices into more than one piece.

The threshold is high enough that any distances big enough to break context will fail to
be considered a duplicate and both entries will be analyzed.